refactor : 자료형 수정 및 불필요한 연관관계 삭제

+++ src/main/java/CDProject/vfmarket/domain/entity/Review.java
@@ -32,7 +32,7 @@ public class Review extends BaseTimeEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double starRate;
+    private Integer starRate;
 
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "item_id")
